# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Restaurant Management System HTTP API.",
        "title": "RMS API",
        "contact": {
            "name": "RMS"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/modifiers/options": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modifier-options"
                ],
                "summary": "List modifier options",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-array_github_com_Jiruu246_rms_internal_dto_ModifierOption"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modifier-options"
                ],
                "summary": "Create a modifier option",
                "parameters": [
                    {
                        "description": "Modifier option details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_internal_dto.CreateModifierOptionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-github_com_Jiruu246_rms_internal_dto_ModifierOption"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/modifiers/options/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modifier-options"
                ],
                "summary": "Get a modifier option by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Modifier option ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-github_com_Jiruu246_rms_internal_dto_ModifierOption"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "modifier-options"
                ],
                "summary": "Delete a modifier option",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Modifier option ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modifier-options"
                ],
                "summary": "Update a modifier option",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Modifier option ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_internal_dto.UpdateModifierOptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-github_com_Jiruu246_rms_internal_dto_ModifierOption"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    }
                }
            }
        },
...
        "/users/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get the current user's profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-github_com_Jiruu246_rms_internal_dto_User"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update the current user's profile",
                "parameters": [
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_internal_dto.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-github_com_Jiruu246_rms_internal_dto_User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIResponse-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_Jiruu246_rms_internal_dto.CreateModifierOptionRequest": {
            "type": "object",
            "required": [
                "modifier_id",
                "name"
            ],
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "image_url": {
                    "type": "string"
                },
                "modifier_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "pre_select": {
                    "type": "boolean"
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
...
        "github_com_Jiruu246_rms_internal_dto.ModifierOption": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "modifier_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pre_select": {
                    "type": "boolean"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_Jiruu246_rms_internal_dto.UpdateModifierOptionRequest": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "image_url": {
                    "type": "string"
                },
                "modifier_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "pre_select": {
                    "type": "boolean"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "github_com_Jiruu246_rms_internal_dto.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_Jiruu246_rms_internal_dto.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Jiruu246_rms_pkg_utils.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_Jiruu246_rms_pkg_utils.APIResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIError"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Jiruu246_rms_pkg_utils.APIResponse-array_github_com_Jiruu246_rms_internal_dto_ModifierOption": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Jiruu246_rms_internal_dto.ModifierOption"
                    }
                },
                "error": {
                    "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIError"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Jiruu246_rms_pkg_utils.APIResponse-github_com_Jiruu246_rms_internal_dto_ModifierOption": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_Jiruu246_rms_internal_dto.ModifierOption"
                },
                "error": {
                    "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIError"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Jiruu246_rms_pkg_utils.APIResponse-github_com_Jiruu246_rms_internal_dto_User": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_Jiruu246_rms_internal_dto.User"
                },
                "error": {
                    "$ref": "#/definitions/github_com_Jiruu246_rms_pkg_utils.APIError"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}